requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
PyJWT
passlib[bcrypt]
black
flake8
autoflake8
pytest
pytest-asyncio
httpx
isort